requirements.txt
numpy
PyOpenGL
PyOpenGL_accelerate
glfw
Pillow
imgui
PIL